change: isolate helper functions in separate module
